+++ src/action.php
@@ -192,6 +192,40 @@ added to ' . $alfred_playlist_name, $track_artwork_path);
 		// Did not find a way to set a timeout
 		$server->run();
 		return;
+	}else if ($type == "NEXT") {
+		exec("osascript -e 'tell application \"Spotify\" to next track'");
+		$tcpport = getFreeTcpPort();
+		exec("osascript -e 'tell application \"Spotify\" to open location \"spotify:app:miniplayer:current_track:" . $tcpport . ":" . uniqid() . "\"'");
+
+		$server = IoServer::factory(
+			new HttpServer(
+				new WsServer(
+					new MiniPlayer()
+				)
+			),
+			$tcpport
+		);
+		// FIX THIS: server will exit when done
+		// Did not find a way to set a timeout
+		$server->run();
+		return;
+	}else if ($type == "PREVIOUS") {
+		exec("osascript -e 'tell application \"Spotify\" to previous track'");
+		$tcpport = getFreeTcpPort();
+		exec("osascript -e 'tell application \"Spotify\" to open location \"spotify:app:miniplayer:current_track:" . $tcpport . ":" . uniqid() . "\"'");
+
+		$server = IoServer::factory(
+			new HttpServer(
+				new WsServer(
+					new MiniPlayer()
+				)
+			),
+			$tcpport
+		);
+		// FIX THIS: server will exit when done
+		// Did not find a way to set a timeout
+		$server->run();
+		return;
 	}
 	
 else if ($type == "ALBUM_OR_PLAYLIST") {
@@ -431,6 +465,40 @@ if ($playlist_uri != "") {
 				unstarCurrentTrack($w);
 				return;
 			} else if ($other_action == "current") {
+				$tcpport = getFreeTcpPort();
+				exec("osascript -e 'tell application \"Spotify\" to open location \"spotify:app:miniplayer:current_track:" . $tcpport . ":" . uniqid() . "\"'");
+		
+				$server = IoServer::factory(
+					new HttpServer(
+						new WsServer(
+							new MiniPlayer()
+						)
+					),
+					$tcpport
+				);
+				// FIX THIS: server will exit when done
+				// Did not find a way to set a timeout
+				$server->run();
+				return;
+			} else if ($other_action == "previous") {
+				exec("osascript -e 'tell application \"Spotify\" to previous track'");
+				$tcpport = getFreeTcpPort();
+				exec("osascript -e 'tell application \"Spotify\" to open location \"spotify:app:miniplayer:current_track:" . $tcpport . ":" . uniqid() . "\"'");
+		
+				$server = IoServer::factory(
+					new HttpServer(
+						new WsServer(
+							new MiniPlayer()
+						)
+					),
+					$tcpport
+				);
+				// FIX THIS: server will exit when done
+				// Did not find a way to set a timeout
+				$server->run();
+				return;
+			} else if ($other_action == "next") {
+				exec("osascript -e 'tell application \"Spotify\" to next track'");
 				$tcpport = getFreeTcpPort();
 				exec("osascript -e 'tell application \"Spotify\" to open location \"spotify:app:miniplayer:current_track:" . $tcpport . ":" . uniqid() . "\"'");